+++ lib/grape-swagger/doc_methods/extensions.rb
@@ -11,6 +11,7 @@ module GrapeSwagger
           add_extension_to(path[method], extension(description)) if description && extended?(description, :x)
 
           settings = route.settings
+          add_extensions_to_operation(settings, path, route) if settings && extended?(settings, :x_operation)
           add_extensions_to_path(settings, path) if settings && extended?(settings, :x_path)
           add_extensions_to_definition(settings, path, definitions) if settings && extended?(settings, :x_def)
         end
@@ -21,6 +22,10 @@ module GrapeSwagger
 
         def add_extensions_to_info(settings, info)
           add_extension_to(info, extension(settings)) if extended?(settings, :x)
+        end
+
+        def add_extensions_to_operation(settings, path, route)
+          add_extension_to(path[route.request_method.downcase.to_sym], extension(settings, :x_operation))
         end
 
         def add_extensions_to_path(settings, path)